config data initialized on installed
