+++ src/main/index.ts
@@ -15,213 +15,9 @@ const isMac = process.platform === 'darwin'
 let mainWindow: BrowserWindow | null = null
 
 //! connect with database start
-import { dbInit } from './db'
+import { dbInit, insertCommandData, insertCommandUtilsData } from './db'
 import { Database } from 'sqlite3'
 const db: Database = dbInit()
-
-// Data to be stored
-const AllCommandConfigure = [
-  {
-    id: `1`,
-    default: true,
-    name: 'Basic Command',
-    command: [`docker`, `run`, `--rm`, `-it`],
-    django: {
-      name: 'django',
-      ports: [7164, 7164]
-    },
-    gazebo: {
-      name: 'gazebo',
-      ports: [6080, 6080]
-    },
-    consoles: {
-      name: 'consoles',
-      ports: [1108, 1108]
-    },
-    other: {
-      name: 'other',
-      ports: [7163, 7163]
-    }
-  },
-  {
-    id: `2`,
-    default: true,
-    name: 'GPU Acceleration Intel',
-    command: [`docker`, `run`, `--rm`, `-it`, `--device`, `/dev/dri`],
-    django: {
-      name: 'django',
-      ports: [7164, 7164]
-    },
-    gazebo: {
-      name: 'gazebo',
-      ports: [6080, 6080]
-    },
-    consoles: {
-      name: 'consoles',
-      ports: [1108, 1108]
-    },
-    other: {
-      name: 'other',
-      ports: [7163, 7163]
-    }
-  },
-  {
-    id: ` 3`,
-    default: true,
-    name: 'GPU Acceleration Nvidia',
-    command: [`docker`, `run`, `--rm`, `-it`, `--gpus`, `all`, `--device`, `/dev/dri`],
-    django: {
-      name: 'django',
-      ports: [7164, 7164]
-    },
-    gazebo: {
-      name: 'gazebo',
-      ports: [6080, 6080]
-    },
-    consoles: {
-      name: 'consoles',
-      ports: [1108, 1108]
-    },
-    other: {
-      name: 'other',
-      ports: [7163, 7163]
-    }
-  },
-  {
-    id: `4`,
-    default: true,
-    name: 'Multiple Gpus',
-    command: [
-      `docker`,
-      `run`,
-      `--rm`,
-      `-it`,
-      `--gpus`,
-      `all`,
-      `--device`,
-      `/dev/dri`,
-      `-e`,
-      `DRI_NAME=card1`
-    ],
-    django: {
-      name: 'django',
-      ports: [7164, 7164]
-    },
-    gazebo: {
-      name: 'gazebo',
-      ports: [6080, 6080]
-    },
-    consoles: {
-      name: 'consoles',
-      ports: [1108, 1108]
-    },
-    other: {
-      name: 'other',
-      ports: [7163, 7163]
-    }
-  },
-  {
-    id: `5`,
-    default: true,
-    name: 'Only Ports',
-    command: [],
-    django: {
-      name: 'django',
-      ports: [7164, 7164]
-    },
-    gazebo: {
-      name: 'gazebo',
-      ports: [6080, 6080]
-    },
-    consoles: {
-      name: 'consoles',
-      ports: [1108, 1108]
-    },
-    other: {
-      name: 'other',
-      ports: [7163, 7163]
-    }
-  },
-  {
-    id: `77`,
-    default: false,
-    name: 'Multiple Gpus - 2',
-    command: [`docker`, `run`, `--rm`, `-it`, `--gpus`, `all`],
-    django: {
-      name: 'django',
-      ports: [7164, 7164]
-    },
-    gazebo: {
-      name: 'gazebo',
-      ports: [6080, 6080]
-    },
-    consoles: {
-      name: 'consoles',
-      ports: [1108, 1108]
-    },
-    other: {
-      name: 'other',
-      ports: [7163, 7163]
-    }
-  }
-]
-
-// Create the table with the adjusted column name
-// is_default BOOLEAN,  -- Renamed from 'default' to 'is_default'
-db.run(
-  `CREATE TABLE IF NOT EXISTS commands (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,  -- Auto-incrementing ID
-    is_default BOOLEAN,
-    name TEXT,
-    command TEXT,
-    django_name TEXT,
-    django_ports TEXT,
-    gazebo_name TEXT,
-    gazebo_ports TEXT,
-    consoles_name TEXT,
-    consoles_ports TEXT,
-    other_name TEXT,
-    other_ports TEXT
-  )`,
-  (err) => {
-    if (err) {
-      console.error('Error creating table: ' + err.message)
-    } else {
-      console.log('Table created or already exists.')
-
-      // Insert data into the table
-      let insertStmt = db.prepare(
-        `INSERT INTO commands (
-        is_default, name, command,
-        django_name, django_ports,
-        gazebo_name, gazebo_ports,
-        consoles_name, consoles_ports,
-        other_name, other_ports
-      ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`
-      )
-
-      AllCommandConfigure.forEach((entry) => {
-        insertStmt.run(
-          entry.default,
-          entry.name,
-          entry.command.join(' '), // Join command array to a single string
-          entry.django.name,
-          entry.django.ports.join(', '), // Join ports array to a single string
-          entry.gazebo.name,
-          entry.gazebo.ports.join(', '),
-          entry.consoles.name,
-          entry.consoles.ports.join(', '),
-          entry.other.name,
-          entry.other.ports.join(', ')
-        )
-      })
-
-      insertStmt.finalize()
-
-      console.log('Data inserted successfully.')
-    }
-  }
-)
 
 //! connect database end
 // splash screen
@@ -320,9 +116,10 @@ const createWindow = (): BrowserWindow => {
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
 
-app.whenReady().then(() => {
+app.whenReady().then(async () => {
   //* Store data
-
+  await insertCommandData(db)
+  await insertCommandUtilsData(db)
   // Modify the origin for all requests to the following urls.
   // Set app user model id for windows
   electronApp.setAppUserModelId('com.roboticsacademy')
